fixes an issue testing the creation of rolebindings
